@@ -1,126 +1,94 @@
 {
-    "name": "@chat21/chat21-ionic",
-    "version": "2.0.12",
-    "author": "Dario De Pascalis - Frontiere21 SRL",
-    "license": "AGPL-3.0",
-    "homepage": "https://www.tiledesk.com",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/chat21/chat21-ionic"
+  "name": "chat-ionic-v4",
+  "version": "3.0.91",
+  "author": "Ionic Framework",
+  "homepage": "https://ionicframework.com/",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/common": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/forms": "~8.2.14",
+    "@angular/http": "^7.2.16",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
+    "@ionic-native/core": "^5.0.7",
+    "@ionic-native/globalization": "^5.28.0",
+    "@ionic-native/keyboard": "^5.28.0",
+    "@ionic-native/splash-screen": "^5.0.0",
+    "@ionic-native/status-bar": "^5.0.0",
+    "@ionic/angular": "^5.0.7",
+    "@ionic/storage": "^2.2.0",
+    "@mdi/font": "^5.1.45",
+    "@ngx-translate/core": "^12.1.2",
+    "@ngx-translate/http-loader": "^4.0.0",
+    "cordova-android": "9.0.0",
+    "cordova-browser": "5.0.4",
+    "cordova-ios": "6.1.1",
+    "cordova-plugin-device": "^2.0.2",
+    "cordova-plugin-ionic-keyboard": "^2.2.0",
+    "cordova-plugin-ionic-webview": "^4.2.1",
+    "cordova-plugin-splashscreen": "^5.0.2",
+    "cordova-plugin-statusbar": "^2.4.2",
+    "cordova-plugin-whitelist": "^1.3.3",
+    "core-js": "^2.5.4",
+    "firebase": "^7.14.1",
+    "material-design-icons": "^3.0.1",
+    "moment": "^2.24.0",
+    "moment-timezone": "^0.5.28",
+    "ngx-perfect-scrollbar": "^9.0.0",
+    "perfect-scrollbar": "^1.5.0",
+    "rxjs": "~6.5.1",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.803.20",
+    "@angular/cli": "~8.3.23",
+    "@angular/compiler": "~8.2.14",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@ionic/angular-toolkit": "^2.1.1",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
+  },
+  "description": "An Ionic project",
+  "cordova": {
+    "plugins": {
+      "cordova-plugin-whitelist": {},
+      "cordova-plugin-statusbar": {},
+      "cordova-plugin-device": {},
+      "cordova-plugin-splashscreen": {},
+      "cordova-plugin-ionic-webview": {
+        "ANDROID_SUPPORT_ANNOTATIONS_VERSION": "27.+"
+      },
+      "cordova-plugin-ionic-keyboard": {}
     },
-    "private": false,
-    "bin": {
-        "chat21-ionic": "./bin/chat21-ionic"
-    },
-    "scripts": {
-        "clean": "ionic-app-scripts clean",
-        "build": "ionic-app-scripts build",
-        "start": "node server.js",
-        "lint": "ionic-app-scripts lint",
-        "ionic:build": "ionic-app-scripts build",
-        "ionic:serve": "ionic-app-scripts serve",
-        "compodoc": "./node_modules/.bin/compodoc -d ./docs/ -p ./tsconfig.json",
-        "postinstall": "ionic cordova platform add browser --save && ionic cordova build browser"
-    },
-    "dependencies": {
-        "@angular/animations": "7.2.9",
-        "@angular/common": "7.2.9",
-        "@angular/compiler": "7.2.9",
-        "@angular/compiler-cli": "7.2.9",
-        "@angular/core": "7.2.9",
-        "@angular/forms": "7.2.9",
-        "@angular/http": "7.2.9",
-        "@angular/platform-browser": "7.2.9",
-        "@angular/platform-browser-dynamic": "7.2.9",
-        "@ionic-native/core": "^5.21.5",
-        "@ionic-native/file": "^5.21.5",
-        "@ionic-native/keyboard": "5.2.0",
-        "@ionic-native/push": "5.2.0",
-        "@ionic-native/splash-screen": "5.2.0",
-        "@ionic-native/sqlite": "5.2.0",
-        "@ionic-native/sqlite-porter": "5.2.0",
-        "@ionic-native/status-bar": "5.2.0",
-        "@ionic/cli-plugin-cordova": "1.6.2",
-        "@ionic/storage": "2.2.0",
-        "@ngx-translate/core": "^11.0.1",
-        "@ngx-translate/http-loader": "^4.0.0",
-        "add": "^2.0.6",
-        "angular-linky": "^1.2.2",
-        "build": "^0.1.4",
-        "cordova-android": "^8.1.0",
-        "cordova-browser": "^6.0.0",
-        "cordova-ios": "^5.1.1",
-        "cordova-osx": "^4.0.2",
-        "cordova-plugin-console": "^1.1.0",
-        "cordova-plugin-device": "^1.1.7",
-        "cordova-plugin-splashscreen": "^4.1.0",
-        "cordova-plugin-statusbar": "^2.4.3",
-        "cordova-plugin-whitelist": "^1.3.4",
-        "cordova-sqlite-storage": "^2.6.0",
-        "device": "^0.3.9",
-        "express": "^4.17.1",
-        "firebase": "^7.12.0",
-        "ionic-angular": "3.9.4",
-        "ionic-plugin-keyboard": "^2.2.1",
-        "ionicons": "4.5.5",
-        "macos": "0.0.1",
-        "material-design-icons": "^3.0.1",
-        "moment": "^2.24.0",
-        "npm": "^6.13.7",
-        "phonegap-plugin-push": "^2.3.0",
-        "promise-polyfill": "8.1.0",
-        "rxjs": "6.4.0",
-        "rxjs-compat": "^6.5.4",
-        "sw-toolbox": "3.6.0",
-        "typescript": "3.3.3333",
-        "uk.co.workingedge.cordova.plugin.sqliteporter": "^1.1.1",
-        "zone.js": "0.8.29"
-    },
-    "resolutions": {
-        "grpc": "^1.12.0"
-    },
-    "devDependencies": {
-        "@angular/cli": "^7.3.9",
-        "@angular/tsc-wrapped": "^4.4.6",
-        "@compodoc/compodoc": "^1.1.11",
-        "@ionic/app-scripts": "^3.2.4",
-        "@ionic/cli-plugin-proxy": "^1.5.9"
-    },
-    "cordovaPlugins": [
-        "cordova-plugin-whitelist",
-        "cordova-plugin-console",
-        "cordova-plugin-device",
-        "cordova-plugin-statusbar",
-        "cordova-plugin-splashscreen",
-        "ionic-plugin-keyboard"
-    ],
-    "cordovaPlatforms": [
-        "ios",
-        {
-            "platform": "ios",
-            "version": "",
-            "locator": "ios"
-        }
-    ],
-    "description": "listaUtenti: An Ionic project",
-    "cordova": {
-        "plugins": {
-            "cordova-sqlite-storage": {},
-            "uk.co.workingedge.cordova.plugin.sqliteporter": {},
-            "cordova-plugin-console": {},
-            "cordova-plugin-device": {},
-            "cordova-plugin-splashscreen": {},
-            "cordova-plugin-statusbar": {},
-            "cordova-plugin-whitelist": {},
-            "ionic-plugin-keyboard": {},
-            "phonegap-plugin-push": {
-                "SENDER_ID": "1096415488178"
-            }
-        },
-        "platforms": [
-            "android",
-            "browser",
-            "osx"
-        ]
-    }
+    "platforms": [
+      "browser",
+      "android",
+      "ios"
+    ]
+  }
 }
